Move load_servant_links() ot lib/servant.py
